Moved Manifest.providers() method to Manifest+Providers.swift file
